Add quick-net for contributing.
This is a project that I've been slowly working on. With the help of some new people, this project should be able to expand. I've outlined 2 issues so far that can be worked on, however I can add more depending on how quickly those are resolved. New contributors are very welcome, no matter the skill level.